feat: refact functions
